CHanging directory name and fixing linter issues
