+++ python-interview-problems-parsing-csv/test_football_v1.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """ Pytest functions for CSV Football problem """
 import pytest
-import football_v1 as fb
+# import football_v1 as fb
 
 
 @pytest.fixture
